requirements.txt
opensearch-py
requests-aws4auth
boto3
python-dotenv